[Keyboard] Updated keyboard & keymaps (#12667)
