+++ keyboards/nack/info.json
@@ -1,7 +1,7 @@
 {
     "keyboard_name": "nack",
-    "url": "https://github.com/pastapojken/nack",
-    "maintainer": "pastapojken",
+    "url": "https://github.com/farfalleflickan/nack",
+    "maintainer": "farfalleflickan",
     "width": 13,
     "height": 4,
     "layouts": {
