--- keyboards/nack/keymaps/pastapojken/readme.md
@@ -1,1 +1,0 @@
-# pastapojken's keymap for nack
