+++ keyboards/nack/mcuconf.h
@@ -28,4 +28,3 @@
 
 #undef STM32_SPI_USE_SPI2
 #define STM32_SPI_USE_SPI2 FALSE
-
