+++ keyboards/nack/nack.h
@@ -1,4 +1,4 @@
-/* Copyright 2020 pastapojken <pastapojken@gmail.com>
+/* Copyright 2020 farfalleflickan <farfalleflickan@gmail.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
